Settings: Fix FC when try to add new apn from UI
The default edit type is CARRIER_EDIT, this make the edit UI can
exit without check the parameter, the invalid record also saved in
database. when try to add a new apn, the invalid record will confict
with it.
use USER_EDITED type instead of it. This will avoid the invalid
record insert in database.

Change-Id: I5863b8c02b55718151b436c1d3de932d45e3bd84
CRs-Fixed: 2190959