Settings: load default configuration when add new APN.
When add a new APN, some configuration of current carrier should
be set in editor UI by default, such as APN protocol, roaming protocol.
it's a chipset PA requirement.
So load the default configuration from telephony layer and set
them to the apn data array when the UI initializes.

Change-Id: Id892e9fb216c3706551e96adf9c0aeef59145975
CRs-Fixed: 2230137