Settings: Fix app notification switch keeps off status after resetting
Turn app's notification switch from 'on' to 'off', then reset all
apps preference, the switch won't turns to 'on' automaticlly.

The importance of the app notification channel is set to 0 when
turns the switch to 'off', and it won't change when resetting the app
perference. So the notification switch keeps 'off' status.

The solution is setting the importance of the app notification channel
to default value when user reset it.

Change-Id: I9aa6011993a0bfc0f4c06808ebf962d04b297763
CRs-Fixed: 2246112