Tethering: Enable band Preference option if 5G is supported.
Currently after APM mode on and off, Country code is unset
and set again. When user is in SoftAP TetherSettings window,
the channel band preference is greyed out and is not being
enabled and hence user is unable to select the band preference.

This commit makes the Tethersettings also listen to
COUNTRY_CODE_CHANGE broadcast and thus enables the band preference
accordingly.

Change-Id: I3728f90556dbe61d8c019d5b6f64a9a30e8cfc10
CRs-Fixed: 2307093