Settings: Show correct AKM on UI when connected
Currently, when AP supports both SAE and WPA2-PSK and user adds
the network for this AP with WPA2-PSK mode, after connect the
security is shown as SAE in the network info dialog window. This
should be WPA2-PSK as the connect config has WPA2-PSK as the
security AKM.

This change updates the UI with the corresponding AKM through
which the connect happened.

Change-Id: I21c9a29f2fbe512f42e2c1c80ccf86d868c0cdd9
CRs-Fixed: 2355707