Settings: Populate correct AKM on UI when connected
Currently, when AP supports both SAE and WPA2-PSK and user adds
the network for this AP with WPA2-PSK mode, after connect the
security is shown as SAE in the network info dialog window. This
should be WPA2-PSK as the connect config has WPA2-PSK as the
security AKM.

This change updates the UI with the corresponding AKM through
which the connect happened. It removes the override which sets
the AKM corresponding to the access point object obtained after
the scanresult.

Change-Id: I0c001e72f34bbb6a9cb7c47cbb2c19c9af8d8c0d
CRs-Fixed: 2409661